invoice_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
Django>=5.0,<6.0
django-cors-headers==4.3.1
reportlab
httpx
python-dotenv
# ... other dependencies
